1. stream.msgLen -> stream.header.MsgLen 2. rtmp stream name with url raw query 3.

+++ pkg/rtmp/client_pull_session.go
@@ -8,7 +8,7 @@ type PullSession struct {
 	*ClientSession
 }
 
-func NewPullSession(obs PullSessionObserver, connectTimeout int64) *PullSession {
+func NewPullSession(obs PullSessionObserver, connectTimeout int) *PullSession {
 	return &PullSession{
 		ClientSession: NewClientSession(CSTPullSession, obs, connectTimeout),
 	}
